+++ packages/headers/src/lib/accept-language.ts
@@ -115,7 +115,7 @@ export class AcceptLanguage implements HeaderValue, Iterable<[string, number]> {
    * @param languages The locale identifiers of the languages to choose from.
    * @returns The most preferred language or `null` if none match.
    */
-  getPreferred(languages: string[]): string | null {
+  getPreferred<Language extends string>(languages: readonly Language[]): Language | null {
     let sorted = languages
       .map((language) => [language, this.getWeight(language)] as const)
       .sort((a, b) => b[1] - a[1])